\name{mm10.rds}
\docType{data}
\alias{mm10.rds}
\alias{mm10 annotation file}
\title{
'mm10' dataset
}
\description{
mm10 annotation file for SCATE
}
\details{
This is a mm10 annotation file for SCATE package.

File sizes: 58 MB.

Source: https://github.com/zji90/SCATE/tree/master/inst/extdata
}
\usage{
}
\examples{
  fl <- system.file("extdata", "mm10.rds", package = "SCATEData")
  a <- readRDS(fl)
}
\keyword{datasets}

